set up schema region test class
