+++ server/src/main/java/org/apache/iotdb/db/metadata/schemaregion/SchemaEngine.java
@@ -226,6 +226,11 @@ public class SchemaEngine {
 
     if (timedForceMLogThread != null) {
       timedForceMLogThread.shutdown();
+      while (true) {
+        if (!timedForceMLogThread.isTerminated()) {
+          break;
+        }
+      }
       timedForceMLogThread = null;
     }
