+++ server/src/test/java/org/apache/iotdb/db/metadata/schemaRegion/SchemaRegionBasicTest.java
@@ -20,49 +20,31 @@ package org.apache.iotdb.db.metadata.schemaRegion;
 
 import org.apache.iotdb.commons.consensus.SchemaRegionId;
 import org.apache.iotdb.commons.exception.MetadataException;
-import org.apache.iotdb.commons.file.SystemFileFactory;
 import org.apache.iotdb.commons.path.MeasurementPath;
 import org.apache.iotdb.commons.path.PartialPath;
 import org.apache.iotdb.commons.path.PathPatternTree;
-import org.apache.iotdb.consensus.ConsensusFactory;
-import org.apache.iotdb.db.conf.IoTDBConfig;
-import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.exception.metadata.AliasAlreadyExistException;
 import org.apache.iotdb.db.exception.metadata.MeasurementAlreadyExistException;
 import org.apache.iotdb.db.exception.metadata.PathAlreadyExistException;
-import org.apache.iotdb.db.metadata.MetadataConstant;
-import org.apache.iotdb.db.metadata.plan.schemaregion.impl.ActivateTemplateInClusterPlanImpl;
 import org.apache.iotdb.db.metadata.plan.schemaregion.impl.CreateTimeSeriesPlanImpl;
-import org.apache.iotdb.db.metadata.plan.schemaregion.impl.DeactivateTemplatePlanImpl;
-import org.apache.iotdb.db.metadata.plan.schemaregion.impl.PreDeactivateTemplatePlanImpl;
-import org.apache.iotdb.db.metadata.plan.schemaregion.impl.RollbackPreDeactivateTemplatePlanImpl;
 import org.apache.iotdb.db.metadata.schemaregion.ISchemaRegion;
 import org.apache.iotdb.db.metadata.schemaregion.SchemaEngine;
-import org.apache.iotdb.db.metadata.template.Template;
-import org.apache.iotdb.db.mpp.plan.statement.metadata.template.CreateSchemaTemplateStatement;
-import org.apache.iotdb.db.qp.physical.sys.CreateTimeSeriesPlan;
-import org.apache.iotdb.db.qp.physical.sys.ShowTimeSeriesPlan;
-import org.apache.iotdb.db.query.dataset.ShowTimeSeriesResult;
 import org.apache.iotdb.tsfile.file.metadata.enums.CompressionType;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSEncoding;
-import org.apache.iotdb.tsfile.utils.Pair;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
-import java.io.File;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -73,142 +55,18 @@ import java.util.stream.IntStream;
  */
 public class SchemaRegionBasicTest extends AbstractSchemaRegionTest {
 
-  IoTDBConfig config = IoTDBDescriptor.getInstance().getConfig();
-  boolean isClusterMode;
-
   public SchemaRegionBasicTest(SchemaRegionTestParams testParams) {
     super(testParams);
   }
 
   @Before
   public void setUp() throws Exception {
-    isClusterMode = config.isClusterMode();
-    config.setClusterMode(true);
     super.setUp();
-    SchemaEngine.getInstance().init();
   }
 
   @After
   public void tearDown() throws Exception {
-    SchemaEngine.getInstance().clear();
     super.tearDown();
-    config.setClusterMode(isClusterMode);
-  }
-
-  @Test
-  public void testRatisModeSnapshot() throws Exception {
-    String schemaRegionConsensusProtocolClass = config.getSchemaRegionConsensusProtocolClass();
-    config.setSchemaRegionConsensusProtocolClass(ConsensusFactory.RATIS_CONSENSUS);
-    try {
-      PartialPath storageGroup = new PartialPath("root.sg");
-      SchemaRegionId schemaRegionId = new SchemaRegionId(0);
-      SchemaEngine.getInstance().createSchemaRegion(storageGroup, schemaRegionId);
-      ISchemaRegion schemaRegion = SchemaEngine.getInstance().getSchemaRegion(schemaRegionId);
-
-      File mLogFile =
-          SystemFileFactory.INSTANCE.getFile(
-              schemaRegion.getStorageGroupFullPath()
-                  + File.separator
-                  + schemaRegion.getSchemaRegionId().getId(),
-              MetadataConstant.METADATA_LOG);
-      Assert.assertFalse(mLogFile.exists());
-
-      Map<String, String> tags = new HashMap<>();
-      tags.put("tag-key", "tag-value");
-      schemaRegion.createTimeseries(
-          new CreateTimeSeriesPlan(
-              new PartialPath("root.sg.d1.s1"),
-              TSDataType.INT32,
-              TSEncoding.PLAIN,
-              CompressionType.UNCOMPRESSED,
-              null,
-              tags,
-              null,
-              null),
-          -1);
-
-      File snapshotDir = new File(config.getSchemaDir() + File.separator + "snapshot");
-      snapshotDir.mkdir();
-      schemaRegion.createSnapshot(snapshotDir);
-
-      schemaRegion.loadSnapshot(snapshotDir);
-
-      Pair<List<ShowTimeSeriesResult>, Integer> result =
-          schemaRegion.showTimeseries(
-              new ShowTimeSeriesPlan(
-                  new PartialPath("root.sg.**"), false, "tag-key", "tag-value", 0, 0, false),
-              null);
-
-      ShowTimeSeriesResult seriesResult = result.left.get(0);
-      Assert.assertEquals(new PartialPath("root.sg.d1.s1").getFullPath(), seriesResult.getName());
-      Map<String, String> resultTagMap = seriesResult.getTag();
-      Assert.assertEquals(1, resultTagMap.size());
-      Assert.assertEquals("tag-value", resultTagMap.get("tag-key"));
-
-      SchemaEngine.getInstance().clear();
-      SchemaEngine.getInstance().init();
-
-      SchemaEngine.getInstance().createSchemaRegion(storageGroup, schemaRegionId);
-      ISchemaRegion newSchemaRegion = SchemaEngine.getInstance().getSchemaRegion(schemaRegionId);
-      newSchemaRegion.loadSnapshot(snapshotDir);
-      result =
-          newSchemaRegion.showTimeseries(
-              new ShowTimeSeriesPlan(
-                  new PartialPath("root.sg.**"), false, "tag-key", "tag-value", 0, 0, false),
-              null);
-
-      seriesResult = result.left.get(0);
-      Assert.assertEquals(new PartialPath("root.sg.d1.s1").getFullPath(), seriesResult.getName());
-      resultTagMap = seriesResult.getTag();
-      Assert.assertEquals(1, resultTagMap.size());
-      Assert.assertEquals("tag-value", resultTagMap.get("tag-key"));
-
-    } finally {
-      config.setSchemaRegionConsensusProtocolClass(schemaRegionConsensusProtocolClass);
-    }
-  }
-
-  @Test
-  @Ignore
-  public void testSnapshotPerformance() throws Exception {
-    String schemaRegionConsensusProtocolClass = config.getSchemaRegionConsensusProtocolClass();
-    config.setSchemaRegionConsensusProtocolClass(ConsensusFactory.RATIS_CONSENSUS);
-    try {
-      PartialPath storageGroup = new PartialPath("root.sg");
-      SchemaRegionId schemaRegionId = new SchemaRegionId(0);
-      SchemaEngine.getInstance().createSchemaRegion(storageGroup, schemaRegionId);
-      ISchemaRegion schemaRegion = SchemaEngine.getInstance().getSchemaRegion(schemaRegionId);
-
-      Map<String, String> tags = new HashMap<>();
-      tags.put("tag-key", "tag-value");
-
-      long time = System.currentTimeMillis();
-      for (int i = 0; i < 1000; i++) {
-        for (int j = 0; j < 1000; j++) {
-          schemaRegion.createTimeseries(
-              new CreateTimeSeriesPlan(
-                  new PartialPath("root.sg.d" + i + ".s" + j),
-                  TSDataType.INT32,
-                  TSEncoding.PLAIN,
-                  CompressionType.UNCOMPRESSED,
-                  null,
-                  tags,
-                  null,
-                  null),
-              -1);
-        }
-      }
-      System.out.println(
-          "Timeseries creation costs " + (System.currentTimeMillis() - time) + "ms.");
-
-      File snapshotDir = new File(config.getSchemaDir() + File.separator + "snapshot");
-      snapshotDir.mkdir();
-      schemaRegion.createSnapshot(snapshotDir);
-
-      schemaRegion.loadSnapshot(snapshotDir);
-    } finally {
-      config.setSchemaRegionConsensusProtocolClass(schemaRegionConsensusProtocolClass);
-    }
   }
 
   @Test
@@ -443,144 +301,5 @@ public class SchemaRegionBasicTest extends AbstractSchemaRegionTest {
         schemaRegion.fetchSchema(new PartialPath("root.**"), Collections.EMPTY_MAP, false);
     Assert.assertEquals(1, schemas.size());
     Assert.assertEquals("root.sg.wf02.wt01.temperature", schemas.get(0).getFullPath());
-  }
-
-  /** Test {@link ISchemaRegion#activateSchemaTemplate}. */
-  @Test
-  public void testActivateSchemaTemplate() throws Exception {
-    PartialPath storageGroup = new PartialPath("root.sg");
-    SchemaRegionId schemaRegionId = new SchemaRegionId(0);
-    SchemaEngine.getInstance().createSchemaRegion(storageGroup, schemaRegionId);
-    ISchemaRegion schemaRegion = SchemaEngine.getInstance().getSchemaRegion(schemaRegionId);
-    schemaRegion.createTimeseries(
-        new CreateTimeSeriesPlanImpl(
-            new PartialPath("root.sg.wf01.wt01.status"),
-            TSDataType.BOOLEAN,
-            TSEncoding.PLAIN,
-            CompressionType.SNAPPY,
-            null,
-            null,
-            null,
-            null),
-        -1);
-    int templateId = 1;
-    Template template =
-        new Template(
-            new CreateSchemaTemplateStatement(
-                "t1",
-                new String[][] {new String[] {"s1"}, new String[] {"s2"}},
-                new TSDataType[][] {
-                  new TSDataType[] {TSDataType.DOUBLE}, new TSDataType[] {TSDataType.INT32}
-                },
-                new TSEncoding[][] {
-                  new TSEncoding[] {TSEncoding.RLE}, new TSEncoding[] {TSEncoding.RLE}
-                },
-                new CompressionType[][] {
-                  new CompressionType[] {CompressionType.SNAPPY},
-                  new CompressionType[] {CompressionType.SNAPPY}
-                }));
-    template.setId(templateId);
-    schemaRegion.activateSchemaTemplate(
-        new ActivateTemplateInClusterPlanImpl(new PartialPath("root.sg.wf01.wt01"), 3, templateId),
-        template);
-    schemaRegion.activateSchemaTemplate(
-        new ActivateTemplateInClusterPlanImpl(new PartialPath("root.sg.wf02"), 2, templateId),
-        template);
-    Set<String> expectedPaths = new HashSet<>(Arrays.asList("root.sg.wf01.wt01", "root.sg.wf02"));
-    Set<String> pathsUsingTemplate =
-        new HashSet<>(schemaRegion.getPathsUsingTemplate(new PartialPath("root.**"), templateId));
-    Assert.assertEquals(expectedPaths, pathsUsingTemplate);
-    PathPatternTree allPatternTree = new PathPatternTree();
-    allPatternTree.appendPathPattern(new PartialPath("root.**"));
-    allPatternTree.constructTree();
-    Assert.assertEquals(2, schemaRegion.countPathsUsingTemplate(templateId, allPatternTree));
-    PathPatternTree wf01PatternTree = new PathPatternTree();
-    wf01PatternTree.appendPathPattern(new PartialPath("root.sg.wf01.*"));
-    wf01PatternTree.constructTree();
-    Assert.assertEquals(1, schemaRegion.countPathsUsingTemplate(templateId, wf01PatternTree));
-    Assert.assertEquals(
-        "root.sg.wf01.wt01",
-        schemaRegion.getPathsUsingTemplate(new PartialPath("root.sg.wf01.*"), templateId).get(0));
-    PathPatternTree wf02PatternTree = new PathPatternTree();
-    wf02PatternTree.appendPathPattern(new PartialPath("root.sg.wf02"));
-    wf02PatternTree.constructTree();
-    Assert.assertEquals(1, schemaRegion.countPathsUsingTemplate(templateId, wf02PatternTree));
-    Assert.assertEquals(
-        "root.sg.wf02",
-        schemaRegion.getPathsUsingTemplate(new PartialPath("root.sg.wf02"), templateId).get(0));
-  }
-
-  /**
-   * Test {@link ISchemaRegion#constructSchemaBlackListWithTemplate}, {@link
-   * ISchemaRegion#rollbackSchemaBlackListWithTemplate} and {@link
-   * ISchemaRegion#deactivateTemplateInBlackList}
-   */
-  @Test
-  public void testDeactivateTemplate() throws Exception {
-    PartialPath storageGroup = new PartialPath("root.sg");
-    SchemaRegionId schemaRegionId = new SchemaRegionId(0);
-    SchemaEngine.getInstance().createSchemaRegion(storageGroup, schemaRegionId);
-    ISchemaRegion schemaRegion = SchemaEngine.getInstance().getSchemaRegion(schemaRegionId);
-    schemaRegion.createTimeseries(
-        new CreateTimeSeriesPlanImpl(
-            new PartialPath("root.sg.wf01.wt01.status"),
-            TSDataType.BOOLEAN,
-            TSEncoding.PLAIN,
-            CompressionType.SNAPPY,
-            null,
-            null,
-            null,
-            null),
-        -1);
-    int templateId = 1;
-    Template template =
-        new Template(
-            new CreateSchemaTemplateStatement(
-                "t1",
-                new String[][] {new String[] {"s1"}, new String[] {"s2"}},
-                new TSDataType[][] {
-                  new TSDataType[] {TSDataType.DOUBLE}, new TSDataType[] {TSDataType.INT32}
-                },
-                new TSEncoding[][] {
-                  new TSEncoding[] {TSEncoding.RLE}, new TSEncoding[] {TSEncoding.RLE}
-                },
-                new CompressionType[][] {
-                  new CompressionType[] {CompressionType.SNAPPY},
-                  new CompressionType[] {CompressionType.SNAPPY}
-                }));
-    template.setId(templateId);
-    schemaRegion.activateSchemaTemplate(
-        new ActivateTemplateInClusterPlanImpl(new PartialPath("root.sg.wf01.wt01"), 3, templateId),
-        template);
-    schemaRegion.activateSchemaTemplate(
-        new ActivateTemplateInClusterPlanImpl(new PartialPath("root.sg.wf02"), 2, templateId),
-        template);
-
-    // construct schema blacklist with template on root.sg.wf01.wt01 and root.sg.wf02
-    Map<PartialPath, List<Integer>> allDeviceTemplateMap = new HashMap<>();
-    allDeviceTemplateMap.put(new PartialPath("root.**"), Collections.singletonList(templateId));
-    schemaRegion.constructSchemaBlackListWithTemplate(
-        new PreDeactivateTemplatePlanImpl(allDeviceTemplateMap));
-
-    // rollback schema blacklist with template on root.sg.wf02
-    Map<PartialPath, List<Integer>> wf02TemplateMap = new HashMap<>();
-    wf02TemplateMap.put(new PartialPath("root.sg.wf02"), Collections.singletonList(templateId));
-    schemaRegion.rollbackSchemaBlackListWithTemplate(
-        new RollbackPreDeactivateTemplatePlanImpl(wf02TemplateMap));
-
-    // deactivate schema blacklist with template on root.sg.wf01.wt01
-    schemaRegion.deactivateTemplateInBlackList(
-        new DeactivateTemplatePlanImpl(allDeviceTemplateMap));
-
-    // check using getPathsUsingTemplate
-    List<String> expectedPaths = Collections.singletonList("root.sg.wf02");
-    List<String> pathsUsingTemplate =
-        schemaRegion.getPathsUsingTemplate(new PartialPath("root.**"), templateId);
-    Assert.assertEquals(expectedPaths, pathsUsingTemplate);
-    // check using countPathsUsingTemplate
-    PathPatternTree allPatternTree = new PathPatternTree();
-    allPatternTree.appendPathPattern(new PartialPath("root.**"));
-    allPatternTree.constructTree();
-    Assert.assertEquals(1, schemaRegion.countPathsUsingTemplate(templateId, allPatternTree));
   }
 }
